fix: cap position loss at margin, correct client money units
The BigInt pnl core is sound - it agrees with a floating-point reference on
every case tested, and loss equals margin exactly at the liquidation price for
all five leverages on both sides. The bugs are around it.

closeOrder marks against the CURRENT price, not the liquidation price. Between
two Binance ticks the price can jump straight past the liquidation level, and
the raw pnl then exceeds the margin: measured on a 10x $100 long, a 20% gap
returns -18000 cents, driving the account $180 negative. A position can never
lose more than its margin - that is what a liquidation level is for - so the
loss is now clamped.

margin is a count of cents and is now validated as an integer. A float put
fractional cents into the balance and every later sum inherited them.

Client, all one root cause - margin is dollars, userBalance is cents:
- the insufficient-balance check compared the two directly, so it was 100x too
  lenient and never fired; the server rejected instead
- the margin fill bar divided dollars by cents and sat at ~0% forever
- the post-trade refresh read balanceResponse.balance.usd_balance, but
  /user/balance returns { usd_balance } with no wrapper, so the guard was
  always false and the displayed balance never updated after a trade

12 tests cover the scales, reference agreement, flat-market zero, the
liquidation invariant, long/short symmetry, and that the gap case really does
exceed the margin without the clamp.

+++ http_server/src/types/userschema.ts
@@ -8,7 +8,9 @@ export const credentailSchma = z.object({
 export const tradeSchema = z.object({
   asset: z.enum(["BTC", "ETH", "SOL"]),
   type: z.enum(["buy", "sell"]),
-  margin: z.number().positive(),
+  // Integer: this is a count of cents. A float here puts fractional cents into
+  // the balance and every downstream sum inherits them.
+  margin: z.number().int().positive(),
   leverage: z.union([
     z.literal(1),
     z.literal(5),
